Restructured the App structure. Created Restaurants with to children components: Filters and List
