+++ src/components/Restaurant/index.js
@@ -1,8 +1,8 @@
 import React from "react";
+import { useParams } from "react-router-dom";
 import { connect } from "react-redux";
 
 import "./Restaurant.scss";
-import { useParams } from "react-router-dom";
 
 const Restaurant = ({ restaurants }) => {
   let { name } = useParams();
